generar.sh: Fix invalid option in Busybox's shuf command
The shuf command provided in the Alpine container is different from the
one provided by the usual (and more powerful) bash shell interpreter,
which you might find in Ubuntu distros.

Busybox's shuf command (the one provided in this Alpine distro) doesn't
have the `--head-count` option, instead we need to use `-n` for the same
purpose. Fix this.

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@gmail.com>